Removes "not in first EXIF" limitation & adds SOF2
The original sizeOfImageForFilePath_JPEG: method could use different segment types to find the size (either EXIF or SOF0, start of frame, baseline DCT). It the size was not found in an EXIF segment, however, it would stop looking instead of continuing to look for a potential SOF0. Also, it appears as though an SOF2, start of frame, progressive DCT should work just like an SOF0 so that was added. Lastly will read-by a Huffman tables segment (0xFFC4) like is does for quantization tables (0xFFDB).